@greatroar let me know if I got this right or not.