@greatroar 2021-11-02
Looks fine from a cursory glance, but I haven't tested this. v4 has better tests, so I only develop there. You could try backporting a few of those to be more sure.