@lizthegrey 2021-11-02
I set this running to 1% of honeycomb production, it seems to be delivering good results.

Before:
![profile001](https://user-images.githubusercontent.com/614704/137598169-253593a6-8088-4603-b8d9-7c9862be9b6f.png)

After:
![profile001](https://user-images.githubusercontent.com/614704/139967404-8baa0761-774f-4d59-b9d6-90a203919ddc.png)

```
File: retriever-lambda
Type: cpu
Time: Nov 2, 2021 at 4:45pm (PDT)
Duration: 359.07s, Total samples = 72.40s (20.16%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) list lz4
Total: 72.40s
ROUTINE ======================== github.com/honeycombio/hound/retriever/cstorage.(*lz4NextSlice).nextSlice in /mnt/ramdisk/hound/retriever/cstorage/varstring_dict.go
         0      1.13s (flat, cum)  1.56% of Total
         .          .    730:	if cap(r.buf) < size {
         .          .    731:		r.buf = make([]byte, size)
         .          .    732:	}
         .          .    733:	res := r.buf[:size]
         .          .    734:
         .      1.13s    735:	_, err := io.ReadFull(r.lz4Reader, res)
         .          .    736:	return res, err
         .          .    737:}
         .          .    738:
         .          .    739:// First read the dictionary, then the remainder of the file looks like any
         .          .    740:// other compact column file, including header.
ROUTINE ======================== github.com/pierrec/lz4/v3.(*Reader).Read in /mnt/ramdisk/go/pkg/mod/github.com/honeycombio/lz4/v3@v3.3.3-0.20211101212407-b7c6134860bb/reader.go
     6.25s     23.18s (flat, cum) 32.01% of Total
         .          .    138:// Read decompresses data from the underlying source into the supplied buffer.
         .          .    139://
         .          .    140:// Since there can be multiple streams concatenated, Header values may
         .          .    141:// change between calls to Read(). If that is the case, no data is actually read from
         .          .    142:// the underlying io.Reader, to allow for potential input buffer resizing.
     566ms      566ms    143:func (z *Reader) Read(buf []byte) (int, error) {
         .          .    144:	if debugFlag {
         .          .    145:		debug("Read buf len=%d", len(buf))
         .          .    146:	}
     1.68s      1.68s    147:	if !z.Header.done {
      12ms      258ms    148:		if err := z.readHeader(true); err != nil {
         .          .    149:			return 0, err
         .          .    150:		}
         .          .    151:		if debugFlag {
         .          .    152:			debug("header read OK compressed buffer %d / %d uncompressed buffer %d : %d index=%d",
         .          .    153:				len(z.zdata), cap(z.zdata), len(z.data), cap(z.data), z.idx)
         .          .    154:		}
         .          .    155:	}
         .          .    156:
     268ms      268ms    157:	if len(buf) == 0 {
         .          .    158:		return 0, nil
         .          .    159:	}
         .          .    160:
      72ms       72ms    161:	if z.idx == len(z.data) {
         .          .    162:		// No data ready for reading, process the next block.
         .          .    163:		if debugFlag {
         .          .    164:			debug("reading block from writer")
         .          .    165:		}
         .          .    166:		// Reset uncompressed buffer
       2ms        2ms    167:		z.data = z.zdata[:cap(z.zdata)][len(z.zdata):]
         .          .    168:
         .          .    169:		// Block length: 0 = end of frame, highest bit set: uncompressed.
         .       52ms    170:		bLen, err := z.readUint32()
         .          .    171:		if err != nil {
         .          .    172:			return 0, err
         .          .    173:		}
         .          .    174:		z.pos += 4
         .          .    175:
         .          .    176:		if bLen == 0 {
         .          .    177:			// End of frame reached.
         .          .    178:			if !z.NoChecksum {
         .          .    179:				// Validate the frame checksum.
         .          .    180:				checksum, err := z.readUint32()
         .          .    181:				if err != nil {
         .          .    182:					return 0, err
         .          .    183:				}
         .          .    184:				if debugFlag {
         .          .    185:					debug("frame checksum got=%x / want=%x", z.checksum.Sum32(), checksum)
         .          .    186:				}
         .          .    187:				z.pos += 4
         .          .    188:				if h := z.checksum.Sum32(); checksum != h {
         .          .    189:					return 0, fmt.Errorf("lz4: invalid frame checksum: got %x; expected %x", h, checksum)
         .          .    190:				}
         .          .    191:			}
         .          .    192:
         .          .    193:			// Get ready for the next concatenated frame and keep the position.
         .          .    194:			pos := z.pos
         .          .    195:			z.Reset(z.src)
         .          .    196:			z.pos = pos
         .          .    197:
         .          .    198:			// Since multiple frames can be concatenated, check for more.
         .          .    199:			return 0, z.readHeader(false)
         .          .    200:		}
         .          .    201:
         .          .    202:		if debugFlag {
         .          .    203:			debug("raw block size %d", bLen)
         .          .    204:		}
         .          .    205:		if bLen&compressedBlockFlag > 0 {
         .          .    206:			// Uncompressed block.
         .          .    207:			bLen &= compressedBlockMask
         .          .    208:			if debugFlag {
         .          .    209:				debug("uncompressed block size %d", bLen)
         .          .    210:			}
         .          .    211:			if int(bLen) > cap(z.data) {
         .          .    212:				return 0, fmt.Errorf("lz4: invalid block size: %d", bLen)
         .          .    213:			}
         .          .    214:			z.data = z.data[:bLen]
         .          .    215:			if _, err := io.ReadFull(z.src, z.data); err != nil {
         .          .    216:				return 0, err
         .          .    217:			}
         .          .    218:			z.pos += int64(bLen)
         .          .    219:			if z.OnBlockDone != nil {
         .          .    220:				z.OnBlockDone(int(bLen))
         .          .    221:			}
         .          .    222:
         .          .    223:			if z.BlockChecksum {
         .          .    224:				checksum, err := z.readUint32()
         .          .    225:				if err != nil {
         .          .    226:					return 0, err
         .          .    227:				}
         .          .    228:				z.pos += 4
         .          .    229:
         .          .    230:				if h := xxh32.ChecksumZero(z.data); h != checksum {
         .          .    231:					return 0, fmt.Errorf("lz4: invalid block checksum: got %x; expected %x", h, checksum)
         .          .    232:				}
         .          .    233:			}
         .          .    234:
         .          .    235:		} else {
         .          .    236:			// Compressed block.
         .          .    237:			if debugFlag {
         .          .    238:				debug("compressed block size %d", bLen)
         .          .    239:			}
         .          .    240:			if int(bLen) > cap(z.data) {
         .          .    241:				return 0, fmt.Errorf("lz4: invalid block size: %d", bLen)
         .          .    242:			}
         .          .    243:			zdata := z.zdata[:bLen]
         .      4.03s    244:			if _, err := io.ReadFull(z.src, zdata); err != nil {
         .          .    245:				return 0, err
         .          .    246:			}
       4ms        4ms    247:			z.pos += int64(bLen)
         .          .    248:
         .          .    249:			if z.BlockChecksum {
         .          .    250:				checksum, err := z.readUint32()
         .          .    251:				if err != nil {
         .          .    252:					return 0, err
         .          .    253:				}
         .          .    254:				z.pos += 4
         .          .    255:
         .          .    256:				if h := xxh32.ChecksumZero(zdata); h != checksum {
         .          .    257:					return 0, fmt.Errorf("lz4: invalid block checksum: got %x; expected %x", h, checksum)
         .          .    258:				}
         .          .    259:			}
         .          .    260:
         .      7.86s    261:			n, err := UncompressBlock(zdata, z.data)
         .          .    262:			if err != nil {
         .          .    263:				return 0, err
         .          .    264:			}
         .          .    265:			z.data = z.data[:n]
         .          .    266:			if z.OnBlockDone != nil {
         .          .    267:				z.OnBlockDone(n)
         .          .    268:			}
         .          .    269:		}
         .          .    270:
         .          .    271:		if !z.NoChecksum {
         .      1.07s    272:			_, _ = z.checksum.Write(z.data)
         .          .    273:			if debugFlag {
         .          .    274:				debug("current frame checksum %x", z.checksum.Sum32())
         .          .    275:			}
         .          .    276:		}
         .          .    277:		z.idx = 0
         .          .    278:	}
         .          .    279:
     750ms      752ms    280:	if z.skip > int64(len(z.data[z.idx:])) {
         .          .    281:		z.skip -= int64(len(z.data[z.idx:]))
         .          .    282:		z.dpos += int64(len(z.data[z.idx:]))
         .          .    283:		z.idx = len(z.data)
         .          .    284:		return 0, nil
         .          .    285:	}
         .          .    286:
     354ms      356ms    287:	z.idx += int(z.skip)
      48ms       48ms    288:	z.dpos += z.skip
      16ms       16ms    289:	z.skip = 0
         .          .    290:
     1.97s      5.64s    291:	n := copy(buf, z.data[z.idx:])
      82ms       84ms    292:	z.idx += n
     282ms      282ms    293:	z.dpos += int64(n)
         .          .    294:	if debugFlag {
         .          .    295:		debug("copied %d bytes to input", n)
         .          .    296:	}
         .          .    297:
     136ms      136ms    298:	return n, nil
         .          .    299:}
         .          .    300:
         .          .    301:// Seek implements io.Seeker, but supports seeking forward from the current
         .          .    302:// position only. Any other seek will return an error. Allows skipping output
         .          .    303:// bytes which aren't needed, which in some scenarios is faster than reading
ROUTINE ======================== github.com/pierrec/lz4/v3.(*Reader).readHeader in /mnt/ramdisk/go/pkg/mod/github.com/honeycombio/lz4/v3@v3.3.3-0.20211101212407-b7c6134860bb/reader.go
         0      246ms (flat, cum)  0.34% of Total
         .          .     42:func (z *Reader) readHeader(first bool) error {
         .          .     43:	defer z.checksum.Reset()
         .          .     44:
         .          .     45:	buf := z.buf[:]
         .          .     46:	for {
         .       70ms     47:		magic, err := z.readUint32()
         .          .     48:		if err != nil {
         .          .     49:			z.pos += 4
         .          .     50:			if !first && err == io.ErrUnexpectedEOF {
         .          .     51:				return io.EOF
         .          .     52:			}
         .          .     53:			return err
         .          .     54:		}
         .          .     55:		if magic == frameMagic {
         .          .     56:			break
         .          .     57:		}
         .          .     58:		if magic>>8 != frameSkipMagic>>8 {
         .          .     59:			return ErrInvalid
         .          .     60:		}
         .          .     61:		skipSize, err := z.readUint32()
         .          .     62:		if err != nil {
         .          .     63:			return err
         .          .     64:		}
         .          .     65:		z.pos += 4
         .          .     66:		m, err := io.CopyN(ioutil.Discard, z.src, int64(skipSize))
         .          .     67:		if err != nil {
         .          .     68:			return err
         .          .     69:		}
         .          .     70:		z.pos += m
         .          .     71:	}
         .          .     72:
         .          .     73:	// Header.
         .        2ms     74:	if _, err := io.ReadFull(z.src, buf[:2]); err != nil {
         .          .     75:		return err
         .          .     76:	}
         .          .     77:	z.pos += 8
         .          .     78:
         .          .     79:	b := buf[0]
         .          .     80:	if v := b >> 6; v != Version {
         .          .     81:		return fmt.Errorf("lz4: invalid version: got %d; expected %d", v, Version)
         .          .     82:	}
         .          .     83:	if b>>5&1 == 0 {
         .          .     84:		return ErrBlockDependency
         .          .     85:	}
         .          .     86:	z.BlockChecksum = b>>4&1 > 0
         .          .     87:	frameSize := b>>3&1 > 0
         .          .     88:	z.NoChecksum = b>>2&1 == 0
         .          .     89:
         .          .     90:	bmsID := buf[1] >> 4 & 0x7
         .          .     91:	if bmsID < 4 || bmsID > 7 {
         .          .     92:		return fmt.Errorf("lz4: invalid block max size ID: %d", bmsID)
         .          .     93:	}
         .          .     94:	bSize := blockSizeIndexToValue(bmsID - 4)
         .          .     95:	z.BlockMaxSize = bSize
         .          .     96:
         .          .     97:	// Allocate the compressed/uncompressed buffers.
         .          .     98:	// The compressed buffer cannot exceed the uncompressed one.
         .          .     99:	if n := 2 * bSize; cap(z.zdata) < n {
         .      170ms    100:		z.zdata = make([]byte, n, n)
         .          .    101:	}
         .          .    102:	if debugFlag {
         .          .    103:		debug("header block max size id=%d size=%d", bmsID, bSize)
         .          .    104:	}
         .          .    105:	z.zdata = z.zdata[:bSize]
         .          .    106:	z.data = z.zdata[:cap(z.zdata)][bSize:]
         .          .    107:	z.idx = len(z.data)
         .          .    108:
         .        2ms    109:	_, _ = z.checksum.Write(buf[0:2])
         .          .    110:
         .          .    111:	if frameSize {
         .          .    112:		buf := buf[:8]
         .        2ms    113:		if _, err := io.ReadFull(z.src, buf); err != nil {
         .          .    114:			return err
         .          .    115:		}
         .          .    116:		z.Size = binary.LittleEndian.Uint64(buf)
         .          .    117:		z.pos += 8
         .          .    118:		_, _ = z.checksum.Write(buf)
ROUTINE ======================== github.com/pierrec/lz4/v3.(*Reader).readUint32 in /mnt/ramdisk/go/pkg/mod/github.com/honeycombio/lz4/v3@v3.3.3-0.20211101212407-b7c6134860bb/reader.go
       2ms      122ms (flat, cum)  0.17% of Total
         .          .    325:	z.checksum.Reset()
         .          .    326:}
         .          .    327:
         .          .    328:// readUint32 reads an uint32 into the supplied buffer.
         .          .    329:// The idea is to make use of the already allocated buffers avoiding additional allocations.
       2ms        2ms    330:func (z *Reader) readUint32() (uint32, error) {
         .          .    331:	buf := z.buf[:4]
         .      120ms    332:	_, err := io.ReadFull(z.src, buf)
         .          .    333:	x := binary.LittleEndian.Uint32(buf)
         .          .    334:	return x, err
         .          .    335:}
ROUTINE ======================== github.com/pierrec/lz4/v3.NewReader in /mnt/ramdisk/go/pkg/mod/github.com/honeycombio/lz4/v3@v3.3.3-0.20211101212407-b7c6134860bb/reader.go
         0        6ms (flat, cum) 0.0083% of Total
         .          .     30:}
         .          .     31:
         .          .     32:// NewReader returns a new LZ4 frame decoder.
         .          .     33:// No access to the underlying io.Reader is performed.
         .          .     34:func NewReader(src io.Reader) *Reader {
         .        6ms     35:	r := &Reader{src: src}
         .          .     36:	return r
         .          .     37:}
         .          .     38:
         .          .     39:// readHeader checks the frame magic number and parses the frame descriptoz.
         .          .     40:// Skippable frames are supported even as a first frame although the LZ4
ROUTINE ======================== github.com/pierrec/lz4/v3.UncompressBlock in /mnt/ramdisk/go/pkg/mod/github.com/honeycombio/lz4/v3@v3.3.3-0.20211101212407-b7c6134860bb/block.go
         0      7.86s (flat, cum) 10.85% of Total
         .          .     25:// An error is returned if the source data is invalid or the destination buffer is too small.
         .          .     26:func UncompressBlock(src, dst []byte) (int, error) {
         .          .     27:	if len(src) == 0 {
         .          .     28:		return 0, nil
         .          .     29:	}
         .      7.86s     30:	if di := decodeBlock(dst, src); di >= 0 {
         .          .     31:		return di, nil
         .          .     32:	}
         .          .     33:	return 0, ErrInvalidSourceShortBuffer
         .          .     34:}
         .          .     35:
ROUTINE ======================== github.com/pierrec/lz4/v3.decodeBlock in /mnt/ramdisk/go/pkg/mod/github.com/honeycombio/lz4/v3@v3.3.3-0.20211101212407-b7c6134860bb/decode_amd64.s
     812ms      838ms (flat, cum)  1.16% of Total
         .          .     36:	MOVQ R9, R13
         .          .     37:	SUBQ $16, R13
         .          .     38:
         .          .     39:loop:
         .          .     40:	// for si < len(src)
      10ms       10ms     41:	CMPQ SI, R9
         .          .     42:	JGE end
         .          .     43:
         .          .     44:	// token := uint32(src[si])
      26ms       26ms     45:	MOVBQZX (SI), DX
       8ms        8ms     46:	INCQ SI
         .          .     47:
         .          .     48:	// lit_len = token >> 4
         .          .     49:	// if lit_len > 0
         .          .     50:	// CX = lit_len
       6ms        6ms     51:	MOVQ DX, CX
       4ms        4ms     52:	SHRQ $4, CX
         .          .     53:
         .          .     54:	// if lit_len != 0xF
      30ms       30ms     55:	CMPQ CX, $0xF
         .          .     56:	JEQ lit_len_loop_pre
      12ms       12ms     57:	CMPQ DI, R12
         .          .     58:	JGE lit_len_loop_pre
       6ms        6ms     59:	CMPQ SI, R13
         .          .     60:	JGE lit_len_loop_pre
         .          .     61:
         .          .     62:	// copy shortcut
         .          .     63:
         .          .     64:	// A two-stage shortcut for the most common case:
         .          .     65:	// 1) If the literal length is 0..14, and there is enough space,
         .          .     66:	// enter the shortcut and copy 16 bytes on behalf of the literals
         .          .     67:	// (in the fast mode, only 8 bytes can be safely copied this way).
         .          .     68:	// 2) Further if the match length is 4..18, copy 18 bytes in a similar
         .          .     69:	// manner; but we ensure that there's enough space in the output for
         .          .     70:	// those 18 bytes earlier, upon entering the shortcut (in other words,
         .          .     71:	// there is a combined check for both stages).
         .          .     72:
         .          .     73:	// copy literal
       4ms        4ms     74:	MOVOU (SI), X0
      24ms       24ms     75:	MOVOU X0, (DI)
      32ms       32ms     76:	ADDQ CX, DI
…